Rework configuration loading in providers
Validation of provider specific configuration is now done in the
provider as it should be.

The internal getProviders function was replaced with getProvider.
This solves the ambiguous ProvidersConfig and ProviderConfigs type
situation by removing the ProviderConfigs type.
It also changes the amount of code that needs to be added for a new
provider from a new case statement with duplicated info to a single
line in a typechecked object.

The type checking will make sure that ProvidersConfig, the interface
that defines the provider's config and generates the JSON schema, and
the allProviders object, that maps a provider's name to the
implementation, remains in sync. The following will be caught:
- Provider in the interface but not in the object and opposite.
- Provider does not have the correct constructor.